Verifica admin/user para funções Musics

+++ src/domains/Music/controller/index.ts
@@ -1,20 +1,27 @@
 import MusicService from '../services/MusicService';
 import statusCodes from '../../../../utils/constants/statusCodes';
 import { Router, Request, Response, NextFunction } from 'express';
+import checkRoles from '../../../middlewares/checkRole';
+import UserRoles from '../../../../utils/constants/userRoles';
+import { verifyJWT } from '../../../middlewares/auth-middlewares';
+import { NotAuthorizedError } from '../../../../errors/NotAuthorizedError';
 
 const router = Router();
 
-router.put('/update', async (req: Request, res: Response, next: NextFunction) => {
+router.put('/update', verifyJWT, checkRoles([UserRoles.ADMIN, UserRoles.USER]), async (req: Request, res: Response, next: NextFunction) => {
 	try {
+		if (req.user.role != UserRoles.ADMIN) {
+			throw new NotAuthorizedError('Você não tem permissão para alterar músicas.');
+		}
 		const { id, artistId, ...rest } = req.body;
 		await MusicService.updateMusic({ id: parseInt(id), artistId: parseInt(artistId), ...rest });
-		res.status(statusCodes.NO_CONTENT).json('Música atualizada com sucesso!');
+		res.status(statusCodes.SUCCESS).json('Música atualizada com sucesso!');
 	} catch (error) {
 		next(error);
 	}
 });
 
-router.get('/allMusics', async (req: Request, res: Response, next: NextFunction) => {
+router.get('/allMusics', verifyJWT, async (req: Request, res: Response, next: NextFunction) => {
 	try {
 		const musics = await MusicService.readAll();
 		res.status(statusCodes.SUCCESS).json(musics);
@@ -24,7 +31,7 @@ router.get('/allMusics', async (req: Request, res: Response, next: NextFunction)
 	}
 });
 
-router.get('/name/:name', async (req: Request, res: Response, next: NextFunction) => {
+router.get('/name/:name', verifyJWT, async (req: Request, res: Response, next: NextFunction) => {
 	try {
 		const musics = await MusicService.readByName(req.params.name);
 		res.status(statusCodes.SUCCESS).json(musics);
@@ -34,7 +41,7 @@ router.get('/name/:name', async (req: Request, res: Response, next: NextFunction
 	}
 });
 
-router.get('/id/:id', async (req: Request, res: Response, next: NextFunction) => {
+router.get('/id/:id', verifyJWT, async (req: Request, res: Response, next: NextFunction) => {
 
 	try {
 		const musics = await MusicService.readById(parseInt(req.params.id));
@@ -45,21 +52,27 @@ router.get('/id/:id', async (req: Request, res: Response, next: NextFunction) =>
 	}
 });
 
-router.post('/create', async (req: Request, res: Response, next: NextFunction) => {
+router.post('/create', verifyJWT, checkRoles([UserRoles.ADMIN, UserRoles.USER]), async (req: Request, res: Response, next: NextFunction) => {
 	try {
+		if (req.user.role != UserRoles.ADMIN) {
+			throw new NotAuthorizedError('Você não tem permissão para criar músicas.');
+		}
 		await MusicService.createMusic(req.body, parseInt(req.body.artistId));
-		res.status(statusCodes.CREATED).json('Musica criada com sucesso!');
+		res.status(statusCodes.CREATED).json('Música criada com sucesso!');
 
 	} catch (error) {
 		next(error);
 	}
 });
 
-router.delete('/delete/:id', async (req: Request, res: Response, next: NextFunction) => {
+router.delete('/delete/:id', verifyJWT, checkRoles([UserRoles.ADMIN, UserRoles.USER]), async (req: Request, res: Response, next: NextFunction) => {
 
 	try {
+		if (req.user.role != UserRoles.ADMIN) {
+			throw new NotAuthorizedError('Você não tem permissão para deletar músicas.');
+		}
 		const musics = await MusicService.deleteMusic(parseInt(req.params.id));
-		res.status(statusCodes.NO_CONTENT).json(musics);
+		res.status(statusCodes.SUCCESS).json('Música deletada com sucesso!');
 	}
 	catch (error) {
 		next(error);
